+++ BilForbruk/BilForbruk.ino
@@ -36,7 +36,7 @@ float distMeasure()
 
 //Genererer batterinivået, mellom 0 og 80
 
-int batteryDrain(int battery) {
+int batteryDrain(byte battery) {
     battery = 80 - (totalDistance/5);
     if (battery < 0) {
         battery = 0;
